[JENKINS-57326] make GitTool abstract to follow JCasC conventions
