+++ src/main/java/org/jenkinsci/plugins/gitclient/JGitTool.java
@@ -1,11 +1,16 @@
 package org.jenkinsci.plugins.gitclient;
 
+import hudson.EnvVars;
 import hudson.Extension;
+import hudson.model.Node;
+import hudson.model.TaskListener;
 import hudson.plugins.git.GitTool;
+import hudson.tools.ToolDescriptor;
 import hudson.tools.ToolProperty;
 import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
 
+import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 
@@ -34,12 +39,14 @@ public class JGitTool extends GitTool {
         this(Collections.<ToolProperty<?>>emptyList());
     }
 
-
-    /** {@inheritDoc} */
-    @Override
-    public GitTool.DescriptorImpl getDescriptor() {
-        return super.getDescriptor();
+    public GitTool forNode(Node node, TaskListener log) throws IOException, InterruptedException {
+        return this;
     }
+
+    public GitTool forEnvironment(EnvVars environment) {
+        return this;
+    }
+
 
     @Extension @Symbol("jgit")
     public static class DescriptorImpl extends GitTool.DescriptorImpl {
